requirements.txt
asgiref==3.9.2
attrs==25.4.0
Django==5.2.6
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
environ==1.0
Faker==37.8.0
gprof2dot==2025.4.14
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==11.3.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
